refactor(core): 3/5. change project structure
