+++ scripts/validate-commit.ts
@@ -1,11 +1,10 @@
 import { Logger } from "../tools/logger";
-import { CommitMsgValidator } from "../tools/process/git/commit-msg";
+import { CommitMsgValidator } from "../tools/process/git/commit-msg-validate";
 
 export async function main(args: any, root: string, logger: Logger): Promise<number> {
-    // tslint:disable-next-line:no-string-literal
-    const engine = new CommitMsgValidator(process.env["GIT_PARAMS"], logger);
+    const engine = new CommitMsgValidator(process.env["HUSKY_GIT_PARAMS"], logger);
 
     const isValid = await engine.validate().toPromise();
 
-    return isValid ? 0 : 11;
+    return isValid ? 0 : 2;
 }
